♻️  Extract thresholds/options into their own files
